Extract viewGlobals into a module
